simplify handling of licenses
Currently, changes to dependencies usually trigger a corresponding need
to run "make generate-go-licenses" to update the LICENSES file.

The ultimate aim of this file is to track that the licenses we use are
compliant with the policies we operate under (i.e. those of the CNCF).
As such, the most important thing is tracking when licenses change.

It isn't helpful for us to have to do busywork just to update the
version number of a dependency in the LICENSES file; that mostly serves
to break dependabot PRs and doesn't provide us with useful
functionality.

Instead, if we log out the module name and the license, we get the
ability to track and audit licenses, providing a useful signal when a
license changes or a new dependency is added that a maintainer should
take a careful look at what changed to ensure compliance.

Signed-off-by: Ashley Davis <ashley.davis@cyberark.com>