improve error message for helm chart version
The following error is common if a repo does not locally contain any
tags:

```console
$ make test-e2e
make/_shared/helm///helm.mk:35: *** helm_chart_version "1bdf96cfeb0fa2-dirty" should start with a "v".  Stop.
```

This is because makefile-modules uses tag-based logic to determine
the chart version. Without tags, it defaults to a hash which fails
the check edited by this commit.

This commit adds a hint about pulling remote tags to help users diagnose
this issue and fix it.

Signed-off-by: Ashley Davis <ashley.davis@cyberark.com>